requirements.txt
transformers
torch
accelerate
sentencepiece
langchain
langchain-community
duckduckgo-search
Pillow
